% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_correlation.R
\name{all_correlation}
\alias{all_correlation}
\title{A Correlation Matrix Function}
\usage{
all_correlation(x)
}
\arguments{
\item{x}{a year from 1950 to 2017.}
}
\description{
This function allows you to get the correlation matrix
for all numeric variables in the data for a year.
}
\examples{
all_correlation(1991)
}
\keyword{correlation}
