% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBA.R
\docType{data}
\name{NBA}
\alias{NBA}
\title{NBA Players stats since 1950}
\format{
A data frame with 24691 rows and 53 variables:
\describe{
\item{Player}{name}
\item{Pos}{Position}
...
}
}
\source{
\url{https://www.kaggle.com/drgilermo/nba-players-stats/data}
}
\usage{
NBA
}
\description{
The data-set contains aggregate individual statistics for 67 NBA seasons.
from basic box-score attributes such as points, assists, rebounds etc.,
to more advanced money-ball like features such as Value Over Replacement.
}
\keyword{datasets}
