% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{A Oldest Player Function}
\usage{
oldest_player(x)
}
\arguments{
\item{x}{a year from 1950 to 2017.}
}
\description{
This function allows you to find the oldest player in the NBA data for that year
(if multiple players have the same age, then return a single player).
}
\examples{
oldest_player(1991)
}
\keyword{age}
