% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scored_most.R
\name{scored_most}
\alias{scored_most}
\title{A Scored Most Player Function}
\usage{
scored_most(x)
}
\arguments{
\item{x}{a year from 1950 to 2017.}
}
\description{
This function allows you to find a scored most player in a certain year.
}
\examples{
scored_most(1991)
}
\keyword{PTS}
\keyword{score}
